move flatbuffer factories
