+++ src/flatbuffers/models/idRegistryEventModel.test.ts
@@ -1,4 +1,4 @@
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { IdRegistryEventType } from '~/flatbuffers/generated/id_registry_event_generated';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import { HubError } from '~/utils/hubErrors';
