+++ src/flatbuffers/models/validations.test.ts
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker';
 import { utils, Wallet } from 'ethers';
 import { arrayify } from 'ethers/lib/utils';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import * as message_generated from '~/flatbuffers/generated/message_generated';
 import MessageModel from '~/flatbuffers/models/messageModel';
 import * as types from '~/flatbuffers/models/types';
