+++ src/rpc/test/eventService.test.ts
@@ -1,6 +1,6 @@
 import { ClientReadableStream } from '@grpc/grpc-js';
 import { utils, Wallet } from 'ethers';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { EventResponse, EventType } from '~/flatbuffers/generated/rpc_generated';
 import IdRegistryEventModel from '~/flatbuffers/models/idRegistryEventModel';
 import MessageModel from '~/flatbuffers/models/messageModel';
