+++ src/storage/engine/index.test.ts
@@ -1,6 +1,6 @@
 import { utils, Wallet } from 'ethers';
 import { err, ok } from 'neverthrow';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { IdRegistryEventType } from '~/flatbuffers/generated/id_registry_event_generated';
 import { CastId, MessageType } from '~/flatbuffers/generated/message_generated';
 import { NameRegistryEventType } from '~/flatbuffers/generated/name_registry_event_generated';
