+++ src/storage/stores/ampStore.test.ts
@@ -1,4 +1,4 @@
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { MessageType, UserId } from '~/flatbuffers/generated/message_generated';
 import MessageModel from '~/flatbuffers/models/messageModel';
 import { AmpAddModel, AmpRemoveModel, UserPostfix } from '~/flatbuffers/models/types';
