+++ src/storage/stores/signerStore.test.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 import { arrayify } from 'ethers/lib/utils';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { IdRegistryEventType } from '~/flatbuffers/generated/id_registry_event_generated';
 import { MessageType } from '~/flatbuffers/generated/message_generated';
 import IdRegistryEventModel from '~/flatbuffers/models/idRegistryEventModel';
