+++ src/storage/stores/userDataStore.test.ts
@@ -1,6 +1,6 @@
 import { utils, Wallet } from 'ethers';
 import { arrayify } from 'ethers/lib/utils';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { UserDataType } from '~/flatbuffers/generated/message_generated';
 import { NameRegistryEventType } from '~/flatbuffers/generated/name_registry_event_generated';
 import IdRegistryEventModel from '~/flatbuffers/models/idRegistryEventModel';
