+++ src/storage/stores/verificationStore.test.ts
@@ -1,5 +1,5 @@
 import { arrayify } from 'ethers/lib/utils';
-import Factories from '~/flatbuffers/factories/flatbuffer';
+import Factories from '~/flatbuffers/factories';
 import { FarcasterNetwork } from '~/flatbuffers/generated/message_generated';
 import MessageModel from '~/flatbuffers/models/messageModel';
 import {
